Alpha  0.6.0
There's still plenty of bugs, there's a few blocks that we could add, and it's slower than we'd like. Unfortunately, if we add more blocks, it will only get slower. Also, the text is smaller than some people would like. That can be fixed in a future update. For now, it works.